NOISSUE - Refactor integration tests: * Group tests by category * Pull out shared methods
